fisk_dimension/requirements.txt
Flask==2.2.2
pyttsx3==2.90
SpeechRecognition==3.8.1
PyCryptodome==3.17
PyJWT==2.6.0
python-dotenv==1.0.1
streamlit==1.36.0
streamlit-mic-recorder==1.16.0
requests==2.28.2
SQLAlchemy==1.4.46
numpy==1.24.2
psycopg2-binary==2.9.9